Remember me cookie values are swapped
The cookie was called rememberMe and within that cookie, the variable was called User. So the effect was a cookie in the browser called rememberMe[User]. This is the opposite of the code in the readme to be inserted into AppController, where it had cookie name "User" and key "rememberMe".